+++ arrow-memory-size-derive/Cargo.toml
@@ -38,3 +38,6 @@ all-features = true
 proc-macro2 = { version = "1.0", default-features = false }
 quote = { version = "1.0", default-features = false }
 syn = { version = "2.0", features = ["extra-traits"] }
+
+[dev-dependencies]
+arrow-memory-size = { workspace = true }
